+++ Kernel/System/Web/InterfaceAgent.pm
@@ -306,8 +306,9 @@ sub Run {
 
         # login is successful
         my %UserData = $UserObject->GetUserData(
-            User  => $User,
-            Valid => 1
+            User          => $User,
+            Valid         => 1,
+            NoOutOfOffice => 1,
         );
 
         # check if the browser supports cookies